refactor(protocol): protocol 境界を adaptor/protocol/ へ一本化 (#1130)
ルート直下 src-tauri/src/protocol/ と adaptor/protocol/ に併存していた
protocol 境界の二重化を解消し、adaptor/protocol/ へ一本化する。

- protocol/ 配下の WsMessage / serialize_message / deserialize_message /
  agent / auth / branch / error / worktree payload type を
  adaptor/protocol/ へ移設
- lib.rs から mod protocol（ルート直下）を削除
- crate::protocol::* import を crate::adaptor::protocol::* へ置換
- roundtrip / serialize-deserialize test を新 module 側へ移設

対外契約（WebSocket message 名・payload JSON shape）と domain / usecase の
振る舞いは変更しない構造リファクタリング。
