+++ src-tauri/src/adaptor/protocol/auth.rs
