+++ src-tauri/src/adaptor/protocol/error.rs
